+++ .nvmrc
@@ -1,1 +1,1 @@
-v19.7.0
+v20.9.0
