+++ package.json
@@ -11,8 +11,9 @@
   "dependencies": {
     "cfenv": "^1.2.4",
     "chart.js": "^4.4.1",
+    "env-helpers": "^1.2.0",
     "mongodb": "^6.3.0",
-    "next": "14.0.3",
+    "next": "14.2.9",
     "pino": "^8.16.2",
     "pretty-ms": "^8.0.0",
     "react": "^18",
